Se agregan logs al remote data source para hacerle seguimiento a la respuesta del api Se modifica el tipo de dato para accesibility a Double Se agrega llamado para obtener una actividad por categoria en el fragment category Se agrega observador para manejar el estado de la respuesta del api Se crea funcion para validar el precio y agregarle el formato solicitado Se cambia titulo del toolbar dependiendo del valor que llega en la categoria

+++ app/src/main/java/com/example/integradorandroid/data/model/TypeActivities.kt
@@ -8,5 +8,5 @@ data class TypeActivities(
     @SerializedName("price") val price : Double,
     @SerializedName("link") val link : String,
     @SerializedName("key") val key : Int,
-    @SerializedName("accessibility") val accessibility : Int
+    @SerializedName("accessibility") val accessibility : Double
 )
